Finished Script 1
Made a variable/argument to be given a number to be used with Nth Prime. cd'ed to NthPrime to then compile it into an executable named NthPrime (gives us the file tree we want). Than inside the file I added ./NthPrime "$ARG" to run the compiled program with the argument given (ex. ./extract_and_compile.sh 17) .